2semester/1exercise/requirements.txt
defusedxml==0.7.1
fonttools==4.39.3
fpdf2==2.7.3
joblib==1.2.0
lxml==4.9.2
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
